+++ tests/functional_small/ops/ops_epetra_multi_vector.cc
@@ -23,9 +23,9 @@ TEST_F(epetraMultiVectorGlobSize15Fixture, multi_vector_clone)
 
 TEST_F(epetraMultiVectorGlobSize15Fixture, multi_vector_extent)
 {
-    ASSERT_TRUE(pressio::ops::extent(*myMv_,0) == numProc_ * localSize_);
-    ASSERT_TRUE(pressio::ops::extent(*myMv_,1) == numVecs_);
-    ASSERT_TRUE(pressio::ops::extent(*myMv_,2) == 1); // check extent over the rank
+    ASSERT_TRUE(pressio::ops::extent(*myMv_,0) == (size_t)numProc_ * localSize_);
+    ASSERT_TRUE(pressio::ops::extent(*myMv_,1) == (size_t)numVecs_);
+    ASSERT_TRUE(pressio::ops::extent(*myMv_,2) == (size_t)1); // check extent over the rank
 }
 
 TEST_F(epetraMultiVectorGlobSize15Fixture, multi_vector_deep_copy)
